+++ packages/components/components/elvis-pagination/src/react/styledComponents.tsx
@@ -4,15 +4,13 @@ import * as ElviaTypography from '@elvia/elvis-typography';
 
 const colors = {
   elviaCharge: ElviaColors.default['primary-colors']['green'].color,
-  elviaOn: ElviaColors.default['primary-colors']['white'].color,
-  elviaOff: ElviaColors.default['primary-colors']['black'].color,
   grey10: ElviaColors.default['grey-colors']['grey-10'].color,
   grey20: ElviaColors.default['grey-colors']['grey-20'].color,
   grey80: ElviaColors.default['grey-colors']['grey-80'].color,
   grey70: ElviaColors.default['grey-colors']['grey-70'].color,
 };
 
-export const typography = {
+const typography = {
   textSm: ElviaTypography.default['text-sm'],
 };