+++ packages/components/components/elvis-accordion/CHANGELOG.md
@@ -4,8 +4,7 @@
 
 ### Bug fix
 
-- Now using elvis-colors pack for colors.
-- Now using elvis-typography pack for typography.
+- Now using elvis-typography package for typography.
 
 ## 2.0.3 (17.11.21)